Put the food class in another file
